% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{fn_simulate_phenotypes}
\alias{fn_simulate_phenotypes}
\title{Simulate phenotype data from purely additive or additive + epistatic effects.
For purely additive effects, all the simulated QTL are used to compute the genetic value of each entry.
For additive + epistatic effects and if the number of effects per network required is less than the number of possible base-3 combinations of effects,
then the first 90\% of the required effects per network corresponding to the first consecutive additive+epistatic effects are included and the remaining are sampled regularly for the remaining effect combinations.
This may result in less than the required effects per network, if the number of combinations are very large.}
\usage{
fn_simulate_phenotypes(
  G,
  n_alleles = 2,
  dist_effects = c("norm", "chi2")[1],
  n_effects = 5,
  purely_additive = TRUE,
  n_networks = 1,
  n_effects_per_network = 50,
  h2 = 0.5,
  pheno_reps = 1,
  verbose = FALSE
)
}
\arguments{
\item{G:}{genotype matrix with $n$ rows, $l x (n_alleles-1)$ columns, and named rows and columns \link{Required}}

\item{n_alleles:}{number of alleles per locus \link{Required}}

\item{dist_effects:}{distribution of genetic effects which may be "norm" for Gaussian or "chi2" for Chi-squared}

\item{n_effects:}{number of additive genetic effects}

\item{purely_additive:}{Simulate only additive effects?}

\item{n_networks:}{number of networks with non-additive effects}

\item{n_effects_per_network:}{maximum number of non-additive effects per network}

\item{h2:}{heritability, i.e., variance due to genetic effects (additive or additive and non-additive) divided by total phenotype variance}

\item{pheno_reps:}{number of times the phenotypes will be recalculated where the error effects are resampled for each sample}

\item{verbose:}{show simulation messages?}
}
\value{
Y: phenotype matrix with $n$ rows, $pheno_reps$ columns, and named rows and columns
b: additive genetic effects, if purely_additive == TRUE, NULL otherwise
E: epistasis matrix, if purely_additive == FALSE, NULL otherwise
b_epi: epistasis effects, if purely_additive == FALSE, NULL otherwise
}
\description{
Simulate phenotype data from purely additive or additive + epistatic effects.
For purely additive effects, all the simulated QTL are used to compute the genetic value of each entry.
For additive + epistatic effects and if the number of effects per network required is less than the number of possible base-3 combinations of effects,
then the first 90\% of the required effects per network corresponding to the first consecutive additive+epistatic effects are included and the remaining are sampled regularly for the remaining effect combinations.
This may result in less than the required effects per network, if the number of combinations are very large.
}
\examples{
G = fn_simulate_genotypes()
list_Y_b_E_b_epi_1 = fn_simulate_phenotypes(G)
list_Y_b_E_b_epi_2 = fn_simulate_phenotypes(G=G, n_alleles=4, dist_effects="chi2", n_effects=25, h2=0.75, pheno_reps=5, verbose=TRUE)
list_Y_b_E_b_epi_3 = fn_simulate_phenotypes(G=G, n_effects=25, purely_additive=FALSE, n_networks=10, n_effects_per_network=50, verbose=TRUE)
}
